Fix for issue #28: EPPlus' AutoFilter bug.
As reported by eobeda, EPPlus has a bug whereby setting the value of
autofilter at all always results in it being true. This fixes that, as
suggested in the bug report, by only setting the property if it needs to
be set to true.

(Also added Roslyn cache directories to .gitignore.)