fixes #38
HeaderHeight is a nullable double on XlsxMediaTypeFormatter, but the
constructor only accepts a non-nullable double. By changing this to nullable
the consumer can now ignore the height and it will use the default. The lack
of this feature only seemed to affect MS Excel. Previously the height was set
to zero, and "hidding" it in MS Excel. Google Spreadsheets seemed unaffected.